refactor: tasks.py -> image.py, skeletonization.py -> skeleton.py
